+++ db/migrate/20120705130038_add_attributes_to_compute_resources.rb
@@ -1,9 +1,9 @@
 class AddAttributesToComputeResources < ActiveRecord::Migration
   def self.up
-    add_column :compute_resources, :attrs, :text
+    add_column :compute_resources, :attrs, :text unless column_exists? :compute_resources, :attrs
   end
 
   def self.down
-    remove_column :compute_resources, :attrs
+    remove_column :compute_resources, :attrs if column_exists? :compute_resources, :attrs
   end
 end
